more clear component organization
