+++ web/src/components/Feature/PipelineCard/styles.module.scss
