+++ web/src/components/Layout/Embeddings/index.tsx
@@ -2,11 +2,12 @@ import { useRouter } from 'next/router';
 import React, { useState, useMemo, useEffect } from 'react';
 
 // Internal imports from the project
-import Highlight from '@/components/ui/highlight';
-import { Input } from '@/components/ui/input';
-import Pagination from '@/components/ui/pagination';
+import Highlight from '@/components/UI/highlight';
+import { Input } from '@/components/UI/input';
+import Pagination from '@/components/UI/pagination';
 import useLogs from '@/hooks/useLogs';
 import { setColor } from '@/lib/utils';
+import { TriangleIcon } from '../Retrievals';
 
 import {
   Table,
@@ -15,9 +16,8 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from '../ui/table';
-
-import Tippy from '@tippyjs/react';
+} from '../../UI/table';
+import { Tooltip } from 'react-tippy';
 
 // Define your dictionary
 const methodDictionary: { [key: string]: string } = {
@@ -29,7 +29,7 @@ const changeMethod = (method: string) => {
   return methodDictionary[method] || method;
 };
 
-export function Retrieval() {
+export function Embeddings() {
   const router = useRouter();
 
   const handleRowClick = (runId: string) => {
@@ -53,11 +53,11 @@ export function Retrieval() {
   const indexOfLastItem = currentPage * itemsPerPage;
   const indexOfFirstItem = indexOfLastItem - itemsPerPage;
   const filteredLogs = logs.filter((log) => {
-    return log.pipelineRunType !== 'embedding';
+    return log.pipelineRunType === 'embedding';
   });
-  const currentItems = logs
+  const currentItems = filteredLogs
     .filter((log) => {
-      return log.pipelineRunType !== 'embedding';
+      return log.pipelineRunType === 'embedding';
     })
     .slice(indexOfFirstItem, indexOfLastItem);
 
@@ -153,18 +153,18 @@ export function Retrieval() {
                     Event
                   </TableHead>
                   <TableHead
-                    className="flex-3"
+                    className="flex-2 w-0"
                     onClick={() => {
-                      setSortField('searchQuery');
+                      setSortField('outcome');
                       setSortDirection(
                         sortDirection === 'asc' ? 'desc' : 'asc'
                       );
                     }}
                   >
-                    Search Query
+                    DocumentID
                   </TableHead>
                   <TableHead
-                    className="flex-3"
+                    className="flex-2"
                     onClick={() => {
                       setSortField('searchResults');
                       setSortDirection(
@@ -172,19 +172,9 @@ export function Retrieval() {
                       );
                     }}
                   >
-                    Search Results
+                    Embedding Chunks
                   </TableHead>
-                  <TableHead
-                    className="flex-4"
-                    onClick={() => {
-                      setSortField('completionResult');
-                      setSortDirection(
-                        sortDirection === 'asc' ? 'desc' : 'asc'
-                      );
-                    }}
-                  >
-                    Completion Result
-                  </TableHead>
+
                   <TableHead
                     className="flex-2 w-0"
                     onClick={() => {
@@ -194,10 +184,11 @@ export function Retrieval() {
                       );
                     }}
                   >
-                    Outcome
+                    Document Length
                   </TableHead>
+
                   <TableHead
-                    className="flex-2 min-w-[30px] flex justify-between items-center"
+                    className="flex-2"
                     onClick={() => {
                       setSortField('score');
                       setSortDirection(
@@ -205,23 +196,7 @@ export function Retrieval() {
                       );
                     }}
                   >
-                    <span>Search Score</span>
-                    <TriangleIcon
-                      className={`transition-transform transform ${sortField === 'score' && sortDirection === 'desc' ? '-rotate-180' : 'rotate-0'} ml-2`}
-                      width="16"
-                      height="16"
-                    />
-                  </TableHead>
-                  <TableHead
-                    className="flex-2"
-                    onClick={() => {
-                      setSortField('evalResults');
-                      setSortDirection(
-                        sortDirection === 'asc' ? 'desc' : 'asc'
-                      );
-                    }}
-                  >
-                    Eval Results
+                    Outcome
                   </TableHead>
                 </TableRow>
               </TableHeader>
@@ -241,54 +216,21 @@ export function Retrieval() {
                       </Highlight>
                     </TableCell>
                     <TableCell>
-                      {log.searchQuery ? truncateText(log.searchQuery, 30) : ''}
-                    </TableCell>
-                    <TableCell>
-                      {log.searchResults && log.searchResults.length > 0
-                        ? truncateText(log.searchResults[0]?.text, 30)
+                      {log.document !== null && log.document.id !== undefined
+                        ? truncateText(log.document.id, 8)
                         : ''}
                     </TableCell>
                     <TableCell>
-                      {log.completionResult
-                        ? truncateText(log.completionResult, 50)
-                        : ''}
+                      {log.embeddingChunks && log.embeddingChunks.length > 0
+                        ? truncateText(log.embeddingChunks, 30)
+                        : 'N/A'}
                     </TableCell>
+                    <TableCell> {log.embeddingChunks.length} </TableCell>
+
                     <TableCell>
                       <Highlight color={setColor(log.outcome)}>
                         {log.outcome === 'success' ? '✓' : '✗'}
                       </Highlight>
-                    </TableCell>
-                    <TableCell>
-                      {log.outcome === 'success' ? log.score : ''}
-                    </TableCell>
-                    <TableCell>
-                      {log.evalResults
-                        ? Object.entries(log.evalResults).map(
-                            ([key, value], i) => (
-                              <div key={key}>
-                                <Tippy
-                                  content={
-                                    <div
-                                      style={{
-                                        width: 400,
-                                        backgroundColor: '#333',
-                                        color: '#fff',
-                                        padding: 10,
-                                        borderRadius: '10px',
-                                      }}
-                                    >
-                                      {value.reason}
-                                    </div>
-                                  }
-                                >
-                                  <span style={{ whiteSpace: 'nowrap' }}>
-                                    <strong>{key}:</strong> {value.score}
-                                  </span>
-                                </Tippy>
-                              </div>
-                            )
-                          )
-                        : 'N/A'}
                     </TableCell>
                   </TableRow>
                 ))}
@@ -324,23 +266,6 @@ function SearchIcon(props: React.SVGProps<SVGSVGElement>) {
     >
       <circle cx="11" cy="11" r="8" />
       <path d="m21 21-4.3-4.3" />
-    </svg>
-  );
-}
-
-export function TriangleIcon(props: React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      {...props}
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M6 9l6 6 6-6" />
     </svg>
   );
 }
