+++ web/src/components/Layout/Footer/index.tsx
@@ -5,7 +5,7 @@ import { FaGithub, FaTwitter, FaRegMoon } from 'react-icons/fa';
 import { FiExternalLink } from 'react-icons/fi';
 
 import styles from './styles.module.scss';
-import { ArrowIcon } from '../ArrowIcon';
+import { ArrowIcon } from '../../icons/ArrowIcon';
 
 function Footer() {
   const [theme, setTheme] = useState('dark');
