+++ web/src/components/Layout/Retrievals/index.tsx
@@ -2,12 +2,11 @@ import { useRouter } from 'next/router';
 import React, { useState, useMemo, useEffect } from 'react';
 
 // Internal imports from the project
-import Highlight from '@/components/ui/highlight';
-import { Input } from '@/components/ui/input';
-import Pagination from '@/components/ui/pagination';
+import Highlight from '@/components/UI/highlight';
+import { Input } from '@/components/UI/input';
+import Pagination from '@/components/UI/pagination';
 import useLogs from '@/hooks/useLogs';
 import { setColor } from '@/lib/utils';
-import { TriangleIcon } from '../Retrievals';
 
 import {
   Table,
@@ -16,8 +15,9 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from '../ui/table';
-import { Tooltip } from 'react-tippy';
+} from '../../UI/table';
+
+import Tippy from '@tippyjs/react';
 
 // Define your dictionary
 const methodDictionary: { [key: string]: string } = {
@@ -29,7 +29,7 @@ const changeMethod = (method: string) => {
   return methodDictionary[method] || method;
 };
 
-export function Embeddings() {
+export function Retrieval() {
   const router = useRouter();
 
   const handleRowClick = (runId: string) => {
@@ -53,11 +53,11 @@ export function Embeddings() {
   const indexOfLastItem = currentPage * itemsPerPage;
   const indexOfFirstItem = indexOfLastItem - itemsPerPage;
   const filteredLogs = logs.filter((log) => {
-    return log.pipelineRunType === 'embedding';
+    return log.pipelineRunType !== 'embedding';
   });
-  const currentItems = filteredLogs
+  const currentItems = logs
     .filter((log) => {
-      return log.pipelineRunType === 'embedding';
+      return log.pipelineRunType !== 'embedding';
     })
     .slice(indexOfFirstItem, indexOfLastItem);
 
@@ -153,18 +153,18 @@ export function Embeddings() {
                     Event
                   </TableHead>
                   <TableHead
-                    className="flex-2 w-0"
+                    className="flex-3"
                     onClick={() => {
-                      setSortField('outcome');
+                      setSortField('searchQuery');
                       setSortDirection(
                         sortDirection === 'asc' ? 'desc' : 'asc'
                       );
                     }}
                   >
-                    DocumentID
+                    Search Query
                   </TableHead>
                   <TableHead
-                    className="flex-2"
+                    className="flex-3"
                     onClick={() => {
                       setSortField('searchResults');
                       setSortDirection(
@@ -172,9 +172,19 @@ export function Embeddings() {
                       );
                     }}
                   >
-                    Embedding Chunks
+                    Search Results
                   </TableHead>
-
+                  <TableHead
+                    className="flex-4"
+                    onClick={() => {
+                      setSortField('completionResult');
+                      setSortDirection(
+                        sortDirection === 'asc' ? 'desc' : 'asc'
+                      );
+                    }}
+                  >
+                    Completion Result
+                  </TableHead>
                   <TableHead
                     className="flex-2 w-0"
                     onClick={() => {
@@ -184,11 +194,10 @@ export function Embeddings() {
                       );
                     }}
                   >
-                    Document Length
+                    Outcome
                   </TableHead>
-
                   <TableHead
-                    className="flex-2"
+                    className="flex-2 min-w-[30px] flex justify-between items-center"
                     onClick={() => {
                       setSortField('score');
                       setSortDirection(
@@ -196,7 +205,23 @@ export function Embeddings() {
                       );
                     }}
                   >
-                    Outcome
+                    <span>Search Score</span>
+                    <TriangleIcon
+                      className={`transition-transform transform ${sortField === 'score' && sortDirection === 'desc' ? '-rotate-180' : 'rotate-0'} ml-2`}
+                      width="16"
+                      height="16"
+                    />
+                  </TableHead>
+                  <TableHead
+                    className="flex-2"
+                    onClick={() => {
+                      setSortField('evalResults');
+                      setSortDirection(
+                        sortDirection === 'asc' ? 'desc' : 'asc'
+                      );
+                    }}
+                  >
+                    Eval Results
                   </TableHead>
                 </TableRow>
               </TableHeader>
@@ -216,21 +241,54 @@ export function Embeddings() {
                       </Highlight>
                     </TableCell>
                     <TableCell>
-                      {log.document !== null && log.document.id !== undefined
-                        ? truncateText(log.document.id, 8)
+                      {log.searchQuery ? truncateText(log.searchQuery, 30) : ''}
+                    </TableCell>
+                    <TableCell>
+                      {log.searchResults && log.searchResults.length > 0
+                        ? truncateText(log.searchResults[0]?.text, 30)
                         : ''}
                     </TableCell>
                     <TableCell>
-                      {log.embeddingChunks && log.embeddingChunks.length > 0
-                        ? truncateText(log.embeddingChunks, 30)
-                        : 'N/A'}
+                      {log.completionResult
+                        ? truncateText(log.completionResult, 50)
+                        : ''}
                     </TableCell>
-                    <TableCell> {log.embeddingChunks.length} </TableCell>
-
                     <TableCell>
                       <Highlight color={setColor(log.outcome)}>
                         {log.outcome === 'success' ? '✓' : '✗'}
                       </Highlight>
+                    </TableCell>
+                    <TableCell>
+                      {log.outcome === 'success' ? log.score : ''}
+                    </TableCell>
+                    <TableCell>
+                      {log.evalResults
+                        ? Object.entries(log.evalResults).map(
+                            ([key, value], i) => (
+                              <div key={key}>
+                                <Tippy
+                                  content={
+                                    <div
+                                      style={{
+                                        width: 400,
+                                        backgroundColor: '#333',
+                                        color: '#fff',
+                                        padding: 10,
+                                        borderRadius: '10px',
+                                      }}
+                                    >
+                                      {value.reason}
+                                    </div>
+                                  }
+                                >
+                                  <span style={{ whiteSpace: 'nowrap' }}>
+                                    <strong>{key}:</strong> {value.score}
+                                  </span>
+                                </Tippy>
+                              </div>
+                            )
+                          )
+                        : 'N/A'}
                     </TableCell>
                   </TableRow>
                 ))}
@@ -266,6 +324,23 @@ function SearchIcon(props: React.SVGProps<SVGSVGElement>) {
     >
       <circle cx="11" cy="11" r="8" />
       <path d="m21 21-4.3-4.3" />
+    </svg>
+  );
+}
+
+export function TriangleIcon(props: React.SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      {...props}
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    >
+      <path d="M6 9l6 6 6-6" />
     </svg>
   );
 }
