+++ web/src/components/Layout/SearchBar/styles.module.scss
