+++ web/src/components/Layout/SubNavigationMenu/index.tsx
@@ -1,22 +1,34 @@
+// External imports
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 
-import styles from './styles.module.scss';
+// Components and styles
 import { NavItemHighlight } from '../NavItemHighlight';
+import styles from './styles.module.scss';
+
+// TypeScript interfaces for state
+interface NavItemHighlightProps {
+  width: number;
+  translateX: number;
+}
+
+interface NavItem {
+  path: string;
+  label: string;
+  width: number;
+  translateX: number;
+}
 
 export function SubNavigationMenu() {
   const [isScrolling, setIsScrolling] = useState<boolean>(false);
   const [navItemHighlightPropsValues, setNavItemHighlightsPropsValues] =
-    useState<{
-      width: number;
-      translateX: number;
-    } | null>(null);
+    useState<NavItemHighlightProps | null>(null);
 
   const router = useRouter();
   const { pipelineId } = router.query;
 
-  const navItems = [
+  const navItems: NavItem[] = [
     { path: '/', label: 'Home', width: 60, translateX: 0 },
     {
       path: `/pipeline/${pipelineId}`,
@@ -28,26 +40,23 @@ export function SubNavigationMenu() {
     { path: '/embeddings', label: 'Embeddings', width: 100, translateX: 230 },
   ];
 
-  // Function to determine active nav item based on current location
-  function getActiveNavItem() {
-    const activeItem = navItems.find((item) => router.pathname === item.path);
-    if (activeItem) {
-      setNavItemHighlightsPropsValues({
-        width: activeItem.width,
-        translateX: activeItem.translateX,
-      });
-    } else {
-      setNavItemHighlightsPropsValues(null); // Reset if no active item is found
-    }
-  }
+  useEffect(() => {
+    const handleScroll = () => setIsScrolling(window.scrollY > 80);
+    window.addEventListener('scroll', handleScroll, false);
+    return () => window.removeEventListener('scroll', handleScroll, false);
+  }, []);
 
   useEffect(() => {
-    getActiveNavItem();
+    const activeItem = navItems.find((item) => router.pathname === item.path);
+    setNavItemHighlightsPropsValues(
+      activeItem
+        ? { width: activeItem.width, translateX: activeItem.translateX }
+        : null
+    );
   }, [router.pathname]);
 
-  function handleHoverNavItem(event: React.MouseEvent<HTMLAnchorElement>) {
-    const navItemElement = event.currentTarget;
-    const itemPath = navItemElement.getAttribute('href');
+  const handleHoverNavItem = (event: React.MouseEvent<HTMLAnchorElement>) => {
+    const itemPath = event.currentTarget.getAttribute('href');
     const hoveredItem = navItems.find((item) => item.path === itemPath);
     if (hoveredItem) {
       setNavItemHighlightsPropsValues({
@@ -55,44 +64,25 @@ export function SubNavigationMenu() {
         translateX: hoveredItem.translateX,
       });
     }
-  }
-
-  function handleLeaveNavItem() {
-    getActiveNavItem();
-  }
-
-  // handle scroll
-  function handleScroll() {
-    if (window.scrollY > 80) {
-      setIsScrolling(true);
-    } else {
-      setIsScrolling(false);
-    }
-  }
-
-  useEffect(() => {
-    window.addEventListener('scroll', handleScroll, false);
-  }, [isScrolling]);
+  };
 
   return (
     <div className={styles.container}>
       <nav
-        onMouseLeave={handleLeaveNavItem}
+        onMouseLeave={() => setNavItemHighlightsPropsValues(null)}
         className={`${styles.subNavigationMenu} ${isScrolling ? styles.scrollingContent : ''}`}
       >
-        {navItemHighlightPropsValues != null ? (
+        {navItemHighlightPropsValues && (
           <NavItemHighlight
             width={navItemHighlightPropsValues.width}
             translateX={navItemHighlightPropsValues.translateX}
           />
-        ) : (
-          ''
         )}
         {navItems.map((item) => (
           <Link
             key={item.path}
             href={item.path}
-            onMouseOver={(event) => handleHoverNavItem(event)}
+            onMouseOver={handleHoverNavItem}
             className={router.pathname === item.path ? styles.selected : ''}
           >
             {item.label}
