+++ web/src/components/Layout/WorkspacesSelect/styles.module.scss
