+++ web/src/components/icons/ArrowIcon/styles.module.scss
