+++ web/src/pages/_app.tsx
@@ -5,7 +5,7 @@ import { PipelineProvider } from '@/context/PipelineContext';
 
 import { useTheme } from 'next-themes';
 
-import { ThemeProvider } from '@/components/ThemeProvider';
+import { ThemeProvider } from '@/components/Containers/ThemeProvider';
 
 import '../styles/globals.css';
