+++ web/src/pages/event/[eventId].tsx
@@ -2,106 +2,103 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 
 import Layout from '@/components/Layout';
-import { Separator } from '@/components/ui/separator';
+import { Separator } from '@/components/UI/separator';
 import useLogs from '@/hooks/useLogs';
 
 import styles from '../../styles/Index.module.scss';
 import { EventSummary, searchResult } from '../../types';
 
 function Component({ eventLog }: { eventLog: EventSummary }) {
-  console.log("eventLog = ", eventLog)
+  console.log('eventLog = ', eventLog);
   return (
     <div className="grid gap-6 lg:gap-8">
       <div className="space-y-4">
-        {eventLog.searchQuery !== "" && <div>
-          <h2 className="pt-2 pb-2 text-lg font-semibold tracking-tight">
-            Query
-          </h2>
-          <div className="pt-2 rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
-            <span className="text-sm font-semibold text-gray-500 select-all sm:text-base/none dark:text-gray-200">
-              {eventLog.searchQuery}
-            </span>
-          </div> 
-        </div>
-        }
-        {eventLog.searchResults.length > 0 && 
-        <div>
-          
-          <h2 className="text-lg pb-2 font-semibold tracking-tight">
-            Search Results
-          </h2>
-          
-          <div className="grid gap-4">
-            {eventLog.searchResults.map(
-              (result: searchResult, index: number) => (
-                <div
-                  key={index}
-                  className="border rounded-lg p-4 bg-gray-50 dark:bg-gray-800"
-                >
-                  <div className="space-y-1">
-                    <h3 className="text-base font-semibold tracking-tight">
-                      {result.text.split('\\n')[0]}
-                    </h3>
-                    <p className="text-sm text-gray-500 dark:text-gray-400">
-                      {result.text.split('\\n').slice(1).join('\n')}
-                    </p>
-                  </div>
-                </div>
-              )
-            )}
+        {eventLog.searchQuery !== '' && (
+          <div>
+            <h2 className="pt-2 pb-2 text-lg font-semibold tracking-tight">
+              Query
+            </h2>
+            <div className="pt-2 rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
+              <span className="text-sm font-semibold text-gray-500 select-all sm:text-base/none dark:text-gray-200">
+                {eventLog.searchQuery}
+              </span>
+            </div>
           </div>
-        </div> 
-        }
-      </div>
-      {eventLog.completionResult !== "N/A" &&
-      <div>
-        <h2 className="text-lg font-semibold tracking-tight">
-          Completion Output
-        </h2>
-        <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
-          <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
-            {/* You can access the completion of your RAG pipeline event queries here. */}
-            {eventLog.completionResult}
-          </span>
-        </div>
-      </div>
-      }
-       {eventLog.document !== null  && 
-       <div>
-        <h2 className="text-lg font-semibold tracking-tight">
-          DocumentID
-        </h2>
-        <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
-          <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
-            {/* You can access the completion of your RAG pipeline event queries here. */}
-            {eventLog.document.id}
-          </span>
-        </div>
-        <h2 className="text-lg font-semibold tracking-tight pt-4">
-          Document Text
-        </h2>
-        <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
-          <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
-            {/* You can access the completion of your RAG pipeline event queries here. */}
-            {eventLog.document.text}
-          </span>
-        </div>
+        )}
+        {eventLog.searchResults.length > 0 && (
+          <div>
+            <h2 className="text-lg pb-2 font-semibold tracking-tight">
+              Search Results
+            </h2>
 
-        <h2 className="text-lg font-semibold tracking-tight pt-4">
-          Metadata
-        </h2>
-        <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
-          <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
-            {/* You can access the completion of your RAG pipeline event queries here. */}
-            {JSON.stringify(eventLog.document.metadata, null, 2)}
-          </span>
-        </div>
-        </div>
-        }
-        {eventLog.embeddingChunks && <div>
-          <h2 className="text-lg font-semibold tracking-tight pt-4">
-            Chunks
+            <div className="grid gap-4">
+              {eventLog.searchResults.map(
+                (result: searchResult, index: number) => (
+                  <div
+                    key={index}
+                    className="border rounded-lg p-4 bg-gray-50 dark:bg-gray-800"
+                  >
+                    <div className="space-y-1">
+                      <h3 className="text-base font-semibold tracking-tight">
+                        {result.text.split('\\n')[0]}
+                      </h3>
+                      <p className="text-sm text-gray-500 dark:text-gray-400">
+                        {result.text.split('\\n').slice(1).join('\n')}
+                      </p>
+                    </div>
+                  </div>
+                )
+              )}
+            </div>
+          </div>
+        )}
+      </div>
+      {eventLog.completionResult !== 'N/A' && (
+        <div>
+          <h2 className="text-lg font-semibold tracking-tight">
+            Completion Output
           </h2>
+          <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
+            <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
+              {/* You can access the completion of your RAG pipeline event queries here. */}
+              {eventLog.completionResult}
+            </span>
+          </div>
+        </div>
+      )}
+      {eventLog.document !== null && (
+        <div>
+          <h2 className="text-lg font-semibold tracking-tight">DocumentID</h2>
+          <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
+            <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
+              {/* You can access the completion of your RAG pipeline event queries here. */}
+              {eventLog.document.id}
+            </span>
+          </div>
+          <h2 className="text-lg font-semibold tracking-tight pt-4">
+            Document Text
+          </h2>
+          <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
+            <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
+              {/* You can access the completion of your RAG pipeline event queries here. */}
+              {eventLog.document.text}
+            </span>
+          </div>
+
+          <h2 className="text-lg font-semibold tracking-tight pt-4">
+            Metadata
+          </h2>
+          <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
+            <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
+              {/* You can access the completion of your RAG pipeline event queries here. */}
+              {JSON.stringify(eventLog.document.metadata, null, 2)}
+            </span>
+          </div>
+        </div>
+      )}
+      {eventLog.embeddingChunks && (
+        <div>
+          <h2 className="text-lg font-semibold tracking-tight pt-4">Chunks</h2>
           <div className="rounded-lg border p-4 bg-gray-50 dark:bg-gray-800">
             <span className="text-sm text-gray-500 select-all sm:text-base/none dark:text-gray-200">
               {/* You can access the completion of your RAG pipeline event queries here. */}
@@ -109,7 +106,7 @@ function Component({ eventLog }: { eventLog: EventSummary }) {
             </span>
           </div>
         </div>
-        }
+      )}
 
       <div>
         {/* <h2 className="text-lg font-semibold tracking-tight">
@@ -157,7 +154,15 @@ export default function EventPage() {
     <Layout>
       <main className={styles.main}>
         <h1 className="text-white text-2xl mb-4">
-          <Link className="pt-1" href={eventLog?.method !== "Embedding" ? "../retrievals" : "../embeddings"} legacyBehavior>
+          <Link
+            className="pt-1"
+            href={
+              eventLog?.method !== 'Embedding'
+                ? '../retrievals'
+                : '../embeddings'
+            }
+            legacyBehavior
+          >
             <a>
               <svg
                 xmlns="http://www.w3.org/2000/svg"
