+++ web/src/pages/login.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/router';
 import { useState } from 'react';
 
 import { createClient } from '@/utils/supabase/component';
-import { Separator } from '@/components/ui/separator';
+import { Separator } from '@/components/UI/separator';
 
 export default function LoginPage() {
   const router = useRouter();
