+++ web/src/pages/providers/integrations.tsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState, lazy, Suspense } from 'react';
 
-import { IntegrationCard } from '@/components/IntegrationCard';
+import { IntegrationCard } from '@/components/Feature/IntegrationCard';
 import Layout from '@/components/Layout';
-import { Separator } from '@/components/ui/separator';
-import LocalProvidersMenu from '@/components/LocalProvidersMenu';
+import { Separator } from '@/components/UI/separator';
+import LocalProvidersMenu from '@/components/Feature/LocalProvidersMenu';
 import { useFetchProviders } from '@/hooks/useFetchProviders';
 import { useModal } from '@/hooks/useModal';
 
-const SecretsModal = lazy(() => import('@/components/SecretsModal'));
+const SecretsModal = lazy(() => import('@/components/Feature/SecretsModal'));
 
 import styles from '@/styles/Index.module.scss';
 import { Provider } from '../../types';
