+++ web/src/pages/settings.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react';
 //import { IntegrationCard } from '@/components/IntegrationCard';
 import Layout from '@/components/Layout';
 //import { PanelHeader } from '@/components/PanelHeader';
-import { Separator } from '@/components/ui/separator';
+import { Separator } from '@/components/UI/separator';
 
 import styles from '../styles/Index.module.scss';
 import { Pipeline } from '../types';
